[ty] Move `class_member` to `member` module
