+++ crates/ty_python_semantic/src/lib.rs
@@ -34,7 +34,6 @@ mod db;
 mod dunder_all;
 pub mod lint;
 pub(crate) mod list;
-mod member;
 mod module_name;
 mod module_resolver;
 mod node_key;
