+++ crates/ty_python_semantic/src/place.rs
@@ -1,7 +1,6 @@
 use ruff_db::files::File;
 
 use crate::dunder_all::dunder_all_names;
-use crate::member::Member;
 use crate::module_resolver::{KnownModule, file_to_module};
 use crate::semantic_index::definition::{Definition, DefinitionState};
 use crate::semantic_index::place::{PlaceExprRef, ScopedPlaceId};
@@ -231,50 +230,6 @@ pub(crate) fn place<'db>(
         RequiresExplicitReExport::No,
         considered_definitions,
     )
-}
-
-/// Infer the public type of a class member/symbol (its type as seen from outside its scope) in the given
-/// `scope`.
-pub(crate) fn class_member<'db>(db: &'db dyn Db, scope: ScopeId<'db>, name: &str) -> Member<'db> {
-    place_table(db, scope)
-        .symbol_id(name)
-        .map(|symbol_id| {
-            let place_and_quals = place_by_id(
-                db,
-                scope,
-                symbol_id.into(),
-                RequiresExplicitReExport::No,
-                ConsideredDefinitions::EndOfScope,
-            );
-
-            if !place_and_quals.place.is_unbound() && !place_and_quals.is_init_var() {
-                // Trust the declared type if we see a class-level declaration
-                return Member::declared(place_and_quals);
-            }
-
-            if let PlaceAndQualifiers {
-                place: Place::Type(ty, _),
-                qualifiers,
-            } = place_and_quals
-            {
-                // Otherwise, we need to check if the symbol has bindings
-                let use_def = use_def_map(db, scope);
-                let bindings = use_def.end_of_scope_symbol_bindings(symbol_id);
-                let inferred = place_from_bindings_impl(db, bindings, RequiresExplicitReExport::No);
-
-                // TODO: we should not need to calculate inferred type second time. This is a temporary
-                // solution until the notion of Boundness and Declaredness is split. See #16036, #16264
-                Member::inferred(match inferred {
-                    Place::Unbound => Place::Unbound.with_qualifiers(qualifiers),
-                    Place::Type(_, boundness) => {
-                        Place::Type(ty, boundness).with_qualifiers(qualifiers)
-                    }
-                })
-            } else {
-                Member::unbound()
-            }
-        })
-        .unwrap_or_default()
 }
 
 /// Infers the public type of an explicit module-global symbol as seen from within the same file.
@@ -701,7 +656,7 @@ fn place_cycle_initial<'db>(
 }
 
 #[salsa::tracked(cycle_fn=place_cycle_recover, cycle_initial=place_cycle_initial, heap_size=ruff_memory_usage::heap_size)]
-fn place_by_id<'db>(
+pub(crate) fn place_by_id<'db>(
     db: &'db dyn Db,
     scope: ScopeId<'db>,
     place_id: ScopedPlaceId,
